services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trekyatra-api"
version = "0.1.0"
description = "FastAPI backend foundation for TrekYatra"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "sqlalchemy>=2.0.35,<3.0.0",
  "alembic>=1.13.2,<2.0.0",
  "psycopg[binary]>=3.2.1,<4.0.0",
  "redis>=5.0.8,<6.0.0",
  "celery>=5.4.0,<6.0.0",
  "httpx>=0.27.0,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "anthropic>=0.40.0,<1.0.0",
  "langchain-core>=0.3.0,<1.0.0",
  "langchain-anthropic>=0.3.0,<1.0.0",
  "langgraph>=0.2.0,<1.0.0",
  "markdown>=3.6,<4.0",
  "email-validator>=2.0.0,<3.0.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]